reporting-service/requirements.txt
Django>=4.2
djangorestframework
python-decouple
requests
mysqlclient

